Update `DisplayAgent` from v1.3 to v1.4

+++ SampleApp/Sources/UI/Views/DisplayViews/Models/Display/DisplayCommonTemplate.swift
@@ -62,7 +62,7 @@ struct DisplayCommonTemplate: Decodable {
             let closeTemplateAfter: Bool?
             
             struct AutoTrigger: Decodable {
-                let delayInMilliseconds: Double
+                let delayInMilliseconds: Int
                 let showTimer: Bool
             }
